Upstream changes to the original wrapper app (AoA via dumpvdl2 #2,
nearest-sample-rate fix #5, systable.conf parse fix #7, README/packaging
updates) are applied to legacy/, where the original app now lives. The
new src/modules/aoa lands as legacy/src/modules/aoa, csv joins
legacy/Cargo.toml, and legacy/Cargo.lock tracks upstream's lock. Root
Cargo.toml/README/packaging keep the rewrite's versions.

+++ legacy/src/modules/session.rs
@@ -24,7 +24,7 @@ pub trait Session {
     async fn on_timeout(&mut self) -> bool;
     async fn get_errors(&mut self) -> String;
 
-    fn get_listening_band(&self) -> &Vec<u16>;
+    fn get_listening_band(&self) -> &Vec<u64>;
 
     async fn end(&mut self, reason: EndSessionReason);
 }
